Working to the next release (#30)
This is a major refactor of the FIRE pipeline and the naming of its outputs and will represent the v0.1.0 of the software.

There are no changes to the underlying methods or ideas, but I have made significant efforts to speed up the software and reduce dependencies while increasing reliability.